refactor(plugins): move AppConfigWrapper under plugins/admin
