+++ public/app/features/plugins/admin/components/AppConfigWrapper.tsx
